Began developing firmware for Arduino to interact with XBee
Device.

Allows for a prototype module that responds to commands and sends back
status reports.

Still need to include ability to chagne meta data, but foundation seems
to be working.

Includes auto retry of sending packets, if it fails for whatever reason,
after 3 tries, it will silenty return to normal state, and record errors
internally.